fix: add example and releserc
