+++ .prettierignore
@@ -1,3 +1,4 @@
 node_modules/
 dist/
 .github/
+example/
