+++ example/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "keycloak-js-mock-example",
+  "homepage": ".",
+  "version": "0.1.0",
+  "private": false,
+  "author": "Ann Bystrova <ann.bystrova96@mail.ru>",
+  "type": "module",
+  "engines": {
+    "node": ">=16",
+    "npm": ">=8"
+  },
+  "scripts": {
+    "start": "ts-node-esm src/index"
+  },
+  "dependencies": {
+    "keycloak-js": "link:../node_modules/keycloak-js",
+    "keycloak-js-mock": "link:.."
+  },
+  "devDependencies": {
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
+  }
+}
