+++ lib/customize.ts
@@ -1,43 +1,5 @@
-import Keycloak, { KeycloakProfile, KeycloakTokenParsed } from 'keycloak-js';
-import { CompactJWSHeaderParameters } from 'jose/dist/types/types';
-import { RecursivePartial } from './types';
-import mergeObject from './mergeObject';
-
-type TPromiseCallback = (instance: Keycloak.KeycloakInstance) => void;
-type TPromiseMethod = 'resolve' | 'reject';
-interface ICustomizePromise {
-  method: TPromiseMethod;
-  resolveCallback?: TPromiseCallback;
-  rejectCallback?: TPromiseCallback;
-}
-export interface ICustomizeDataPromises {
-  init: ICustomizePromise;
-  logout: ICustomizePromise;
-  updateToken: ICustomizePromise;
-  login: ICustomizePromise;
-  register: ICustomizePromise;
-  accountManagement: ICustomizePromise;
-  loadUserProfile: ICustomizePromise;
-  loadUserInfo: ICustomizePromise;
-}
-interface ICustomizeDataJwt {
-  alg: CompactJWSHeaderParameters['alg'];
-  expiration: number | string;
-  secret: string;
-}
-type TCustomizeUrls = 'login' | 'logout' | 'register' | 'account';
-type TCustomizeDataUrl = {
-  [K in TCustomizeUrls]: string;
-};
-
-export interface ICustomizeData {
-  profile: KeycloakProfile;
-  jwt: ICustomizeDataJwt;
-  promises: Partial<ICustomizeDataPromises>;
-  userInfo: Record<string, unknown>;
-  url: TCustomizeDataUrl;
-  tokenParsed: KeycloakTokenParsed;
-}
+import { ICustomize, ICustomizeData } from './types';
+import { mergeObject } from './utils';
 
 export const initialData: Readonly<ICustomizeData> = {
   profile: {
@@ -88,16 +50,7 @@ export const initialData: Readonly<ICustomizeData> = {
 
 export const customizeData: ICustomizeData = { ...initialData };
 
-export interface ICustomize {
-  profile(data: KeycloakProfile): void;
-  userInfo(data: Record<string, unknown>): void;
-  jwt(data: Partial<ICustomizeDataJwt>): void;
-  url(data: Partial<TCustomizeDataUrl>): void;
-  tokenParsed(data: RecursivePartial<KeycloakTokenParsed>): void;
-  promises(data: RecursivePartial<ICustomizeDataPromises>): void;
-  reset(): void;
-}
-const customize: ICustomize = {
+export const customize: ICustomize = {
   profile(data) {
     customizeData.profile = { ...customizeData.profile, ...data };
   },
@@ -123,5 +76,3 @@ const customize: ICustomize = {
     });
   },
 };
-
-export default customize;
