+++ lib/types/index.ts
@@ -1,3 +1,6 @@
+export * from './customize.types';
+export * from './token.types';
+
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type Merge<A, B> = A & B extends infer U extends Object ? { [K in keyof U]: U[K] } : never;
