+++ lib/types/token.types.ts
@@ -1,0 +1,8 @@
+import {JWTPayload} from "jose";
+
+export interface IGenerateToken {
+  data: JWTPayload;
+  alg: string;
+  expiration: number | string;
+  secret: string;
+}
