+++ lib/utils/keycloakPromise.ts
@@ -1,6 +1,6 @@
 import { KeycloakPromise, KeycloakPromiseCallback } from 'keycloak-js';
 
-function createPromise<TSuccess, TError>(fn: Promise<TSuccess>): { promise: KeycloakPromise<TSuccess, TError> } {
+export function createPromise<TSuccess, TError>(fn: Promise<TSuccess>): { promise: KeycloakPromise<TSuccess, TError> } {
   // @ts-ignore
   const promise: KeycloakPromise<TSuccess, TError> = new Promise<TSuccess>(function (resolve, reject) {
     Promise.resolve(fn)
@@ -27,5 +27,3 @@ function createPromise<TSuccess, TError>(fn: Promise<TSuccess>): { promise: Keyc
   };
   return { promise };
 }
-
-export default createPromise;
